refactor imports, separate user types into two files, and remove temporary get user id function
