+++ backend/src/document-utils/updateDocumentMetadata.ts
@@ -1,10 +1,11 @@
 // This file contains functions that will update the various metadata fields of a Document
-import { ShareStyle } from "@lib/documentTypes";
-import { processDocumentUpdate } from "./documentOperations";
-import FirebaseWrapper from "../firebase-utils/FirebaseWrapper";
-
 import firebase from "firebase/compat/app";
 import "firebase/compat/firestore";
+
+import { ShareStyle } from "@lib/documentTypes";
+
+import { processDocumentUpdate } from "./documentOperations";
+import FirebaseWrapper from "../firebase-utils/FirebaseWrapper";
 
 export async function updateDocumentShareStyle(
   documentId: string,
